requirements.txt
Flask==2.3.3
tensorflow>=2.16.0
numpy>=1.24.0
opencv-python>=4.8.0
pandas>=2.0.0
Pillow>=10.0.0
Werkzeug>=2.3.0
gunicorn>=21.0.0
